+++ src/main/java/calculator/Program.java
@@ -7,11 +7,11 @@ public class Program {
     public static void main(String[] args) {
         System.out.println("Введите поочередно два целых числа и знак операции (‘+’, ‘-’, ‘*’ или ‘/’): ");
         Scanner scanner = new Scanner(System.in);
-        int x = 0, y = 0;
+        double x = 0, y = 0;
         String calc = "";
         try {
-            x = scanner.nextInt();
-            y = scanner.nextInt();
+            x = scanner.nextDouble();
+            y = scanner.nextDouble();
             calc = scanner.next();
         } catch (InputMismatchException ime) {
             System.out.println("Неверные данные, попробуйте снова.");
